+++ src/aosm/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 ++++++++
 * added sns generate-config command
 * added a check to make sure resource type used in ARM template are in allowed list
+* Fixed multi NF RETs issue in nsdvs
 2.0.0b1
 ++++++++
 * Renamed nfdvName to nfdv in CGVs