+++ src/aosm/HISTORY.rst
@@ -4,9 +4,9 @@ Release History
 ===============
 2.0.0b2
 ++++++++
-* Added check to make sure resource type used in ARM template are in allowed list
+* added sns generate-config command
+* added a check to make sure resource type used in ARM template are in allowed list
 * Fixed multi NF RETs issue in nsdvs
-
 2.0.0b1
 ++++++++
 * Renamed nfdvName to nfdv in CGVs